% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{abs_pct_error}
\alias{abs_pct_error}
\title{Absolute Percentage Error}
\usage{
abs_pct_error(x, y)
}
\arguments{
\item{x}{[numeric()] [vector()] actual values}

\item{y}{[numeric()] [vector()] forecast values}
}
\value{
[numeric()] [vector()] absolute percentage errors
}
\description{
Absolute Percentage Error
}
\examples{
x <- seq(0, 10, 0.1)
y <- seq(10, 0, -0.1)
m <- abs_pct_error(x, y)
plot(m, type = "l")

}
