% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{age11_res}
\alias{age11_res}
\alias{age12_res}
\title{Example MVE results to create a summary figure for the lags that were
selected}
\format{
[list()] as output from [multiview_embedding()].

An object of class \code{list} of length 10.
}
\source{
Generated from data-raw/age11.R
}
\usage{
age11_res

age12_res
}
\description{
A list object containing the results of running [multiview_embedding()],
specifically for the Columbia River Bonneville stock for the SalmonPrize
competition in 2026, just the `age11` and `age12` results, with the latter
being more extensive.
}
\keyword{datasets}
