% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{binary}
\alias{binary}
\title{Represent An Integer As A Binary Vector}
\usage{
binary(x, digits = NULL)
}
\arguments{
\item{x}{[integer()] >= 0}

\item{digits}{[integer()] length of the output vector}
}
\value{
[integer()] [vector()] binary representation of x
}
\description{
Represent An Integer As A Binary Vector
}
\examples{
binary(10, digits = 8)

for (i in 0:32) print(binary(i, digits = 6))

}
