% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{check_rEDM_excl_3}
\alias{check_rEDM_excl_3}
\title{Prediction results from setting `exclusionRadius` to 3 in `rEDM:Simplex()`}
\format{
Vector of predicted values from setting `exclusionRadius` to 3.
}
\source{
Generated from Andy running `data-raw/exclusion_radius_test.R`.
}
\usage{
check_rEDM_excl_3
}
\description{
See inclusion issue vignette.
}
\keyword{datasets}
