% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{combine_lag_tibbles}
\alias{combine_lag_tibbles}
\title{Make and Combind Tibbles of Lagged Columns}
\usage{
combine_lag_tibbles(data, names, dims, lag, init = 0)
}
\arguments{
\item{data}{A data frame with a named numeric column to be lagged}

\item{names}{The name of the column to be lagged (character scalar)}

\item{dims}{The number of lagged columns (numeric vector)}

\item{lag}{Number of time steps separating successive lags (numeric scalar)}

\item{init}{The size of the first lag, usually 0 (numeric scalar)}
}
\value{
A tibble of lagged columns
}
\description{
Make and Combind Tibbles of Lagged Columns
}
\examples{
dat <- data.frame(x = 1:15, y = 11:25)
combine_lag_tibbles(dat, c("x", "y"), c(3, 2), 1)

}
