% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lags.R
\name{create_lags}
\alias{create_lags}
\title{Create Lags Of A Vector Or Matrix}
\usage{
create_lags(x, n = 1L)
}
\arguments{
\item{x}{[vector()] or column [matrix()] to lag}

\item{n}{[integer()] lag sizes}
}
\value{
[vector()] or column [matrix()] of lagged values
}
\description{
Create Lags Of A Vector Or Matrix
}
\examples{
create_lags(1:10, 3)
create_lags(matrix(rep(1:10, 2), nrow = 10), 3)
create_lags(matrix(rep(1:10, 2), nrow = 10), c(3, 5))

create_lags(matrix(rep(1:10, 2), nrow = 10), c(0, 1))
create_lags(matrix(rep(1:10, 2), nrow = 10), c(0, 0))
create_lags(matrix(rep(1:10, 2), nrow = 10), c(0, -1))

}
\author{
Luke A. Rogers
}
