% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_subset_lags.R
\name{create_subset_lags}
\alias{create_subset_lags}
\title{Create Subset Lags}
\usage{
create_subset_lags(lags)
}
\arguments{
\item{lags}{[list()] of a named vector of lags for each explanatory
variable.}
}
\value{
[list()] with one lag [list()] for each subset state space
  reconstruction
}
\description{
Creates all combinations, though not in quite the logical order one might
  expect. Could resolve by reording the code maybe, but it works (Andy
  manually checked a detailed example, essentially the one given).
}
\examples{
create_subset_lags(list(a = c(0, 1, 2), b = c(0, 1)))

}
