% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{f}
\alias{f}
\title{Saturating Functional Response}
\usage{
f(u, a, b)
}
\arguments{
\item{u}{[numeric()] State variable value.}

\item{a}{[numeric()] Parameter a_i.}

\item{b}{[numeric()] Parameter b_i.}
}
\value{
[numeric()]
}
\description{
Saturating Functional Response
}
