% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{gets3dusr}
\alias{gets3dusr}
\title{Obtain axes ranges (adapted from function by Uwe Ligges)}
\usage{
gets3dusr(s3dobject)
}
\arguments{
\item{s3dobject}{scatterplot3d plot presumably}
}
\value{
Ranges of x, y and z axes
}
\description{
Obtain axes ranges for scatterplot3d plot (and maybe does others), adapted
from http://r.789695.n4.nabble.com/Axis-Limits-in-Scatterplot3d-td892026.html<description>
}
\examples{
  s3d <- scatterplot3d::scatterplot3d(rnorm(5), rnorm(5), rnorm(5))
  gets3dusr(s3d)
}
\author{
Andrew Edwards (adapted from Uwe Ligges' function)
}
