% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{larkin}
\alias{larkin}
\title{Simulated Larkin Sockeye Salmon Population Dynamics}
\format{
[data.frame()] with 500 rows (brood years) and four variables:
\describe{
  \item{R_3}{Recruitment abundance at age three}
  \item{R_4}{Recruitment abundance at age four}
  \item{R_5}{Recruitment abundance at age five}
  \item{S}{Spawning abundance}
}
}
\source{
Generated from data-raw/larkin.R
}
\usage{
larkin
}
\description{
A dataset containing sockeye salmon population dynamics simulated
from a Larkin spawner-recruitment model using the \code{larkin} package.
}
\keyword{datasets}
