% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{mae}
\alias{mae}
\title{Mean Absolute Error}
\usage{
mae(x, y)
}
\arguments{
\item{x}{[numeric()] [vector()] actual values}

\item{y}{[numeric()] [vector()] forecast values}
}
\value{
[numeric()] mean absolute error
}
\description{
Mean Absolute Error
}
\examples{
mae(1:10, 10:1)

}
