% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{make_dist_tibble}
\alias{make_dist_tibble}
\title{Make a Tibble of Euclidean Distances Between Rows of a Matrix}
\usage{
make_dist_tibble(mat)
}
\arguments{
\item{mat}{A matrix of row vectors (matrix)}
}
\value{
A tibble of Euclidean distances
}
\description{
Make a Tibble of Euclidean Distances Between Rows of a Matrix
}
\examples{
make_dist_tibble(matrix(c(1:6, NA, 8:24), nrow = 6))

}
