% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_previous.R
\name{make_global_indices}
\alias{make_global_indices}
\title{Make Global Allowable Indices from a Data Frame of Lagged Column Row Vectors}
\usage{
make_global_indices(lag_tibble, from = seq_len(nrow(lag_tibble)),
  dist = 1L)
}
\arguments{
\item{lag_tibble}{A data frame of lagged columns defining row vectors}

\item{from}{A vector of indices to consider forecasting from (numeric)}

\item{dist}{The forecast distance (numeric scalar)}

\item{into}{A vector of indices to consider forecasting into (numeric)}
}
\value{
A tibble with two columns of indices
}
\description{
Make Global Allowable Indices from a Data Frame of Lagged Column Row Vectors
}
\examples{
dat <- data.frame(x = c(1:7, NA, 9:20), y = 21:40)
lag_tibble <- make_lag_tibble(dat, list(x = 0:2, y = 0:1))
make_global_indices(lag_tibble)

}
