% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{make_lag_tibble}
\alias{make_lag_tibble}
\title{Make a Tibble of Lagged Columns from a Vector}
\usage{
make_lag_tibble(data, name, dim, lag, init = 0)
}
\arguments{
\item{data}{A data frame with a named numeric column to be lagged}

\item{name}{The name of the column to be lagged (character scalar)}

\item{dim}{The number of lagged columns (numeric scalar)}

\item{lag}{Number of time steps separating successive lags (numeric scalar)}

\item{init}{The size of the first lag, usually 0 (numeric scalar)}
}
\value{
A tibble of successively lagged columns
}
\description{
Rows with NAs are replaced by rows of NAs.
}
\examples{
make_lag_tibble(data.frame(x = 1:16), "x", 3, 2)

}
