% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{make_local_indices}
\alias{make_local_indices}
\title{Make Local Allowable Indices for Nearest Neighbour Vectors}
\usage{
make_local_indices(index, dim, from, lag = 1, dist = 1, symm = FALSE)
}
\arguments{
\item{index}{Index for the focal vector (numeric scalar)}

\item{dim}{Number of embedding dimensions (numeric scalar)}

\item{from}{A vector of indices to consider forecasting from (numeric)}

\item{lag}{Number of time steps separating successive lags (numeric scalar)}

\item{dist}{The forecast distance (numeric scalar)}

\item{symm}{Symmetric exclusion radius? (logical scalar)}
}
\value{
A vector of allowable indices for nearest neighbours
}
\description{
Make Local Allowable Indices for Nearest Neighbour Vectors
}
\examples{
make_local_indices(8, 3, 1:15)

}
