% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-mve-spaces-tibble.R
\name{make_mve_spaces_tibble}
\alias{make_mve_spaces_tibble}
\title{Make a tibble from MVE results to then use in a plot, indicating the axes used in the top multiview embedding reconstructions}
\usage{
make_mve_spaces_tibble(res_mve, ...)
}
\arguments{
\item{res_mve}{list object resulting from a call to [multiview_embedding()]}

\item{...}{passed onto [create_named_lags()], namely `order` to order the
variables consistently.}
}
\value{
tibble of top MVE subsets
}
\description{
Takes the results list from [multiview_embedding()] and indicates the
variables and lags used in the top reconstruction subsets, with the first
column being the respective rho value.
}
\examples{
\dontrun{
  make_mve_spaces_tibble(age11_res)
}

}
