% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{make_neighbours}
\alias{make_neighbours}
\title{Make a Tibble of Nearest Neighbour Indices}
\usage{
make_neighbours(from_index, from_local, distance_tibble, max_neighbours,
  forecast_distance = 1L)
}
\arguments{
\item{from_index}{The time index of the focal vector}

\item{from_local}{The allowable neighbour indices (numeric vector)}

\item{max_neighbours}{The maximum number of neighbours}

\item{forecast_distance}{The forecast distance (numeric scalar)}

\item{dist_tibble}{A tibble with columns 'focal', 'nbr' and 'distance'}
}
\value{
A tibble
}
\description{
Make a Tibble of Nearest Neighbour Indices
}
\examples{
    dist_tibble <- tibble::tibble(x = 1:100) \%>\% 
    make_lag_tibble(list(x = 0:4)) \%>\% 
    make_dist_tibble()
    make_neighbours(50, c(6:49, 56:99), dist_tibble, 6)

}
