% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{matric}
\alias{matric}
\title{Metric Wrapper}
\usage{
matric(x, fun, ...)
}
\arguments{
\item{x}{[numeric()][matrix()] of two columns}

\item{fun}{[function()] taking two [numeric()][vector()] arguments and
returning one scalar [numeric()] value}

\item{...}{additional arguments to pass to \code{fun()}}
}
\value{
[numeric()] scalar
}
\description{
Metric Wrapper
}
\examples{
x <- rnorm(100, 0, 1)
x1 <- matrix(c(x, x), ncol = 2)
matric(x1, mase)

x2 <- rbind(x1, matrix(c(NA_real_, NA_real_), ncol = 2))
matric(x2, mase)

}
