% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{mpe}
\alias{mpe}
\title{Mean Percent Error}
\usage{
mpe(x, y)
}
\arguments{
\item{x}{[numeric()] [vector()] actual values}

\item{y}{[numeric()] [vector()] forecast values}
}
\value{
[numeric()] mean percent error
}
\description{
Mean Percent Error
}
\examples{
mpe(1:10, 10:1)

}
