% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{mre}
\alias{mre}
\title{Mean Raw Error}
\usage{
mre(x, y)
}
\arguments{
\item{x}{[numeric()][vector()] observed values}

\item{y}{[numeric()][vector()] forecast values}
}
\value{
[numeric()] scalar
}
\description{
Mean Raw Error
}
