% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiview_embedding_original.R
\name{multiview_embedding_original}
\alias{multiview_embedding_original}
\title{Multiview Embedding (Luke's original function).}
\usage{
multiview_embedding_original(
  data,
  response,
  lags,
  index = 50L,
  buffer = 10L,
  window = integer(0),
  metric = "rmse",
  beyond = FALSE,
  weight = NULL,
  n_weight = ceiling(sqrt(2^length(unlist(lags)))),
  cores = NULL
)
}
\arguments{
\item{data}{[matrix()] or [data.frame()] with named [numeric()] columns}

\item{response}{[character()] column name of the response variable in
\code{data}}

\item{lags}{[list()] of a named vector of lags for each explanatory
variable.}

\item{index}{[integer()]}

\item{buffer}{[integer()] number of forecasts prior to \code{index}}

\item{window}{[integer()] forecast metric moving window width}

\item{metric}{[character()]}

\item{beyond}{[logical()]}

\item{weight}{TBD}

\item{n_weight}{[integer()]}

\item{cores}{[integer()] number of cores for parallel computation.}
}
\value{
[list()]
}
\description{
See Andy's `multiview_embedding()` for what he's using now.
}
\details{
From Luke's email (TODO tidy up when done):
Good to hear from you! Yes, what you describe is the main idea behind
multiview embedding (MVE). It uses the `2^N - 1` subsets of the superset
state space reconstruction of covariates and covariate lags, giving all
possible embedding dimensions and embedding combinations that you're looking
for. [Andy: Ye and Sugihara 2016 exclude those that have no 0 lag variables,
since presumably they're just shifted - keep in for now and check results
agree maybe].
Andy asked: is single-view embedding just standard EDM, and Luke said:
I think MVE was originally a generalization of S-mapping [Andy: no, for MVE
they just use the single nearest neighbour in each view], but we've
implemented it as a generalization of EDM. So in our case, single-view
embedding would be EDM, but other authors might hear 'single-view embedding'
and think 'S-mapping'.
Worth double-checking because this is just from memory ;)
 Andy: TODO don't think that's quite right, checking
the code.
TODO TODO TODO check the definitions; stick with mve for now to get it all
working.
}
\author{
Luke A. Rogers
}
