% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mve.R
\name{mve}
\alias{mve}
\title{Multi-View Embedding}
\usage{
mve(
  data,
  response,
  lags,
  index = 50L,
  buffer = 10L,
  window = integer(0),
  metric = "rmse",
  beyond = FALSE,
  weight = NULL,
  n_weight = ceiling(sqrt(length(unlist(lags)))),
  cores = NULL
)
}
\arguments{
\item{data}{[matrix()] or [data.frame()] with named [numeric()] columns}

\item{response}{[character()] column name of the response variable in
\code{data}}

\item{lags}{[list()] of a named vector of lags for each explanatory
variable.}

\item{index}{[integer()]}

\item{buffer}{[integer()] number of forecasts prior to \code{index}}

\item{window}{[integer()] forecast metric moving window width}

\item{metric}{[character()]}

\item{beyond}{[logical()]}

\item{weight}{TBD}

\item{n_weight}{[integer()]}

\item{cores}{[integer()] number of cores for parallel computation.}
}
\value{
[list()]
}
\description{
Multi-View Embedding
}
\author{
Luke A. Rogers
}
