% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NY_lags_example}
\alias{NY_lags_example}
\title{Summary results from a simple simulation model analysed using `rEDM` and Andy's
original independent code}
\format{
A 100x14 tibble where each row is a time step (year) and columns
  are:
  * `t`: time step, 1, 2, 3, ..., 100.
  * `N_t`: original time series of number of spawners, $N_t$ at time $t$ from
  Carrie's `salmonTraj()` function.
  * `N_tmin1`: $N_{t-1}$, the number at the previous time step (value for
     `t=1` is therefore `NA`).
  * `Y_t`: First difference $Y_t = N_{t+1} - N_t$; (value for `t=100` is
  therefore `NA`).
  * `Ytmin1`: $Y_{t-1}$
  * `Ytmin2`: $Y_{t-2}$
  * `rEDM.pred`: predicted value from `rEDM::simplex()` with embedding
  dimension $E=2$.
  * `rEDM.var`: estimated variance of `rEDM.pred` (from `rEDM::simplex()`).
  * `my.pred`: predicted value from Andy's `EDM_pred_E_2()` function (saved
  in `PBSedm`)
  * `my.var`: estimated variance of `my.pred`
  * `pred.diff`: the difference between predictions, i.e. `rEDM.pred - my.pred`
  * `var.diff`: the difference between estimated variances, i.e. `rEDM.var - my.var`
  * `pred.ratio`: ratio of the two estimated predictions, as `rEDM.pred / my.pred`
  * `var.ratio`: ratio of the two estimated variances, as `rEDM.var / my.var`
}
\source{
Generated from Andy running `data-raw/Nx_lags.R`.
}
\usage{
NY_lags_example
}
\description{
Contains the original times series of populations `N_t`, simple lagged values
and differences, and then calculations from `rEDM::simplex()` and Andy's
original independent code. This was independent code originally done to understand
EDM, but revealed discrepancies between those results and those of `rEDM()`.
See `full_calcs_orig` and `psi_orig` for full results.
}
\keyword{datasets}
