% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{pbs_calc_dist}
\alias{pbs_calc_dist}
\title{Calculate Distances Among Tibble Row Vectors}
\usage{
pbs_calc_dist(row_tbl)
}
\arguments{
\item{row_tbl}{A tibble of row vectors}
}
\value{
A long tibble of distances between pairs of row vectors
}
\description{
Calculate Distances Among Tibble Row Vectors
}
\examples{
pbs_calc_dist(pbs_make_lags(tibble::tibble(x = 1:100), "x", 3, 1))

}
