% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbs_edm.R
\name{pbs_edm}
\alias{pbs_edm}
\title{Title}
\usage{
pbs_edm(data_frame, lags, forecast_distance = 1L,
  symmetric_exclusion = FALSE, include_stats = TRUE,
  include_forecasts = TRUE, include_neighbours = TRUE)
}
\arguments{
\item{data_frame}{}

\item{lags}{}

\item{forecast_distance}{}

\item{symmetric_exclusion}{}

\item{include_stats}{}

\item{include_forecasts}{}

\item{include_neighbours}{}

\item{interpret_lags}{}

\item{from_user}{}

\item{into_user}{}
}
\description{
Title
}
\examples{
data_frame <- data.frame(x = simple_ts)
lags <- list(x = 0:1)
pbs_edm(data_frame, lags)

}
