% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_previous.R
\name{pbs_make_lags}
\alias{pbs_make_lags}
\title{Create a Tibble of Lagged Copies of a Named Tibble Column}
\usage{
pbs_make_lags(tbl, col_name, embed_dim, lag_size)
}
\arguments{
\item{tbl}{A tibble with a named column to be lagged}

\item{col_name}{The name of a column in tbl}

\item{embed_dim}{An integer embedding dimension corresponding to the number
of lagged columns to create}

\item{lag_size}{The number of time steps separating successive lags}
}
\value{
A tibble of successively lagged columns
}
\description{
Create a Tibble of Lagged Copies of a Named Tibble Column
}
\examples{
pbs_make_lags(tibble::tibble(x = 1:15), "x", 5, 1)

}
