% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{pbs_make_nbrs}
\alias{pbs_make_nbrs}
\title{Make a Tibble of Nearest Neighbours}
\usage{
pbs_make_nbrs(lag_dist, time_ind, rel_lib_ind, embed_dim, pred_dist,
  max_nbrs = embed_dim + 1)
}
\arguments{
\item{lag_dist}{Distances between neighbours (tibble with columns:
'focal_ind', 'nbr_ind', and 'distance')}

\item{time_ind}{The time index of the focal vector (numeric scalar)}

\item{rel_lib_ind}{The allowable neighbour indices (numeric vector)}

\item{embed_dim}{The embedding dimension (numeric scalar)}

\item{pred_dist}{The prediction distance (numeric scalar)}

\item{max_nbrs}{The maximum number of neighbours (numeric scalar). It
should be the embed_dim + 1 for pbs_simplex() and length(rel_lib_ind)
for pbs_s_map().}
}
\value{
An tibble of ordered neighbour indices
}
\description{
Make a Tibble of Nearest Neighbours
}
\examples{
1:100 \%>\% pbs_make_tibble("x") \%>\% 
    pbs_make_lags("x", 5, 1) \%>\% pbs_calc_dist() \%>\%
    pbs_make_nbrs(50, c(6:49, 56:99), 5, 1, 6)

}
