% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{pbs_make_tibble}
\alias{pbs_make_tibble}
\title{Attempt to Return a Tibble with a Named Column}
\usage{
pbs_make_tibble(data, col_name)
}
\arguments{
\item{data}{An object to be returned as a tibble (a tibble, vector with 
null dimensions, or object coercable by tibble::as_tibble)}

\item{col_name}{A name to identify or apply to a column (character)}
}
\value{
A tibble with a named column
}
\description{
Attempt to Return a Tibble with a Named Column
}
\examples{
pbs_make_tibble(1:10, "x")

}
