% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbs_previous.R
\name{pbs_s_map_1}
\alias{pbs_s_map_1}
\title{Use An S-Map to Perform Out-of-Sample Forecasting}
\usage{
pbs_s_map_1(
  data,
  col_name = "x",
  embed_dim = 2L,
  local_weight = 0L,
  lag_size = 1L,
  pred_dist = 1L,
  one_sided = TRUE,
  lib_ind = seq_len(NROW(data)),
  pred_ind = seq_len(NROW(data))
)
}
\arguments{
\item{data}{A time series with possible NA values (numeric vector or
tibble with a named column).}

\item{col_name}{If `data` is a tibble, the time series column name 
(character scalar).}

\item{embed_dim}{The embedding dimension for the lagged-coordinate 
embedding of the time series (integer scalar)}

\item{local_weight}{The exponential local weighting parameter
(integer scalar)}

\item{lag_size}{The number of time steps separating successive lags
(integer scalar)}

\item{pred_dist}{The number of time steps forward to forecast
(integer scalar)}

\item{one_sided}{Should only the focal value be excluded? (logical)}

\item{lib_ind}{Time series indices to predict from (integer vector).
Indices will be removed if the associated lagged coordinate vector 
or its projection forward by pred_dist (1) has one or more
NA values, (2) contains one or more indices that appear in the focal
vector, or (3) falls outside the time series.}

\item{pred_ind}{Time series indices to predict to (integer vector).
Indices will be removed if the associated lagged coordinate vector 
or its projection forward by pred_dist (1) has one or more
NA values, (2) contains one or more indices that appear in the focal
vector, or (3) falls outside the time series.}
}
\value{
A list of tibbles
}
\description{
Use An S-Map to Perform Out-of-Sample Forecasting
}
\examples{
pbs_s_map_1(simple_ts, "x", local_weight = 2)

}
