% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbs_simplex.R
\name{pbs_simplex}
\alias{pbs_simplex}
\title{Perform out-of-sample forecasting via simplex projection}
\usage{
pbs_simplex(time_series, embed_dim, libr_ind = seq_along(time_series),
  pred_ind = libr_ind, lag_size = 1, forecast_dist = 1)
}
\arguments{
\item{time_series}{A numeric vector with possible NA values}

\item{embed_dim}{An integer embedding dimension}

\item{libr_ind}{An integer vector of indices to predict from}

\item{pred_ind}{An integer vector of indicies to predict to}

\item{lag_size}{The number of time steps separating successive lags}

\item{forecast_dist}{The number of time steps forward for each forecast}
}
\description{
Perform out-of-sample forecasting via simplex projection
}
