% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbs_smap.R
\name{pbs_smap}
\alias{pbs_smap}
\title{Use S-Mapping to Perform Out-of-Sample Forecasting}
\usage{
pbs_smap(
  data_frame,
  lags,
  local_weight = 0,
  forecast_distance = 1L,
  first_difference = FALSE,
  centre_and_scale = FALSE,
  show_calculations = FALSE
)
}
\arguments{
\item{data_frame}{A data frame with named columns (list of numeric vectors)}

\item{lags}{Named list of numeric vectors giving lags to use for each
column. List names must match column names. (list of numeric vectors)}

\item{local_weight}{State-dependence parameter (numeric scalar)}

\item{forecast_distance}{Forecast distance (numeric scalar)}

\item{first_difference}{Replace Nt by Xt = Nt - Nt+1? (logical)}

\item{centre_and_scale}{Replace Nt by Xt = (Nt - mean(Nt)) / sd(Nt)? 
(logical)}

\item{show_calculations}{Logical. FALSE: return a tibble of results; TRUE:
return a list including a tibble of results, and calculations.}
}
\value{
A tibble with results, forecasts, neighbours, distances, and weights
}
\description{
Use S-Mapping to Perform Out-of-Sample Forecasting
}
\examples{
data_frame <- data.frame(x = simple_ts)
lags <- list(x = 0:1)
pbs_smap(data_frame, lags)

}
