% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbs_previous.R
\name{pbs_smap_1}
\alias{pbs_smap_1}
\title{Use S-Mapping to Perform Out-of-Sample Forecasting}
\usage{
pbs_smap_1(data_frame, lags, local_weight = 0,
  from_user = seq_len(nrow(data_frame)),
  into_user = seq_len(nrow(data_frame)), forecast_distance = 1L,
  symmetric_exclusion = FALSE, include_stats = TRUE,
  include_forecasts = TRUE, include_neighbours = TRUE)
}
\arguments{
\item{data_frame}{A data frame with named columns (list of numeric vectors)}

\item{lags}{Named list of numeric vectors giving lags to use for each
column. List names must match column names. (list of numeric vectors)}

\item{local_weight}{State-dependence parameter (numeric scalar)}

\item{from_user}{Rows to predict from (numeric vector)}

\item{into_user}{Rows to predict into (numeric vector)}

\item{forecast_distance}{Forecast distance (numeric scalar)}

\item{symmetric_exclusion}{Symmetric exclusion radius? (logical scalar)}

\item{include_stats}{Return forecast stats? (logical scalar)}

\item{include_forecasts}{Return forecast and observations? (logical scalar)}

\item{include_neighbours}{Return neighbours? (logical scalar)}
}
\value{
A tibble (work in progress)
}
\description{
Use S-Mapping to Perform Out-of-Sample Forecasting
}
\examples{
  data_frame <- data.frame(x = simple_ts)
  lags <- list(x = 0:3)
  local_weight <- 0
  pbs_smap_1(data_frame, lags, local_weight)

}
