% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{pbsEDM_Evec}
\alias{pbsEDM_Evec}
\title{Do the `pbsEDM()` calculation for vector of `E` values}
\usage{
pbsEDM_Evec(Nt, E_vec = 1:10, ...)
}
\arguments{
\item{Nt}{[vector] Vector of non-differenced values N(t), with time assumed
to be `1:length(Nt)`.}

\item{E_vec}{The vector of embedding dimensions to try.}
}
\value{
List of `pbsEDM` lists, each main component corresponds to a value
  of `E`, given by `results$E`
}
\description{
Only works for one variable, non-differenced N(t), for now, as I haven't thought how to incorporate
the lags variable properly yet.
}
\examples{
# pbsEDM_Evec(Nx_lags_orig$Nt)
}
\author{
Andrew Edwards
}
