% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbsEDM_optimal_E.R
\name{pbsEDM_optimal_E}
\alias{pbsEDM_optimal_E}
\title{Do the `pbsEDM()` calculation for vector of `E` values and then return the
optimal `E`.}
\usage{
pbsEDM_optimal_E(N_t, lags_to_test = 1:10, ...)
}
\arguments{
\item{N_t}{[vector] Vector of non-differenced values $N_t$, with time assumed
to be `1:length(N_t)`.}

\item{lags_to_test}{vec Vector of lags to try (gets passed onto `E_vec` in
`pbsEDM_Evec`; see Issue 28).}

\item{...}{Further options to pass to `pbsEDM_Evec()` and then onto `pbsEDM()`.}
}
\value{
A `pbsEDM` list corresponding to the optimal `E`, based on `N_rho`.
}
\description{
Only works for one variable, non-differenced $N_t$, for now, as I haven't thought how to incorporate
the lags variable properly yet. Calls `pbsEDM_Evec` and then just returns
the results for the optimal `E`, based on highest `rho`.
}
\examples{
\donttest{
  aa <- pbsEDM_optimal_E(NY_lags_example$N_t)
}
}
\author{
Andrew Edwards
}
