% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{pbsLAG}
\alias{pbsLAG}
\title{Return a Lagged Matrix}
\usage{
pbsLAG(x, n = 1)
}
\arguments{
\item{x}{[matrix()] A vector or column matrix}

\item{n}{[integer()] The lag sizes}
}
\value{
A matrix or vector. If `x` is a vector then returns a vector `length(x)`
  with `NA` for the first `n` values then the first `length(x) - n` values of
  `x`.
If `x` is a matrix then TODO [Andy thinks it does the
  obvious thing].
}
\description{
Return a Lagged Matrix
}
\examples{
pbsLAG(1:10, 3)
pbsLAG(matrix(rep(1:10, 2), nrow = 10), 3)
pbsLAG(matrix(rep(1:10, 2), nrow = 10), c(3, 5))

}
