% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{pbsSim}
\alias{pbsSim}
\title{Simulate Three Species Dynamics via Hastings & Powell Model}
\usage{
pbsSim(
  length_out,
  time_step = 1,
  burn_steps = 10,
  x0 = 0.8,
  y0 = 0.2,
  z0 = 9,
  a1 = 5,
  a2 = 0.1,
  b1 = 2,
  b2 = 2,
  d1 = 0.4,
  d2 = 0.01
)
}
\arguments{
\item{length_out}{[integer()] Time series length}

\item{time_step}{[numeric()] Sample frequency relative to t = R0 * T}

\item{burn_steps}{[integer()] Number of burn in steps}

\item{x0}{[numeric()] Predator state variable.}

\item{y0}{[numeric()] Prey state variable.}

\item{z0}{[numeric()] Producer state variable.}

\item{a1}{[numeric()] Parameter a1}

\item{a2}{[numeric()] Parameter a2}

\item{b1}{[numeric()] Parameter b1}

\item{b2}{[numeric()] Parameter b2}

\item{d1}{[numeric()] Parameter d1}

\item{d2}{[numeric()] Parameter d2}
}
\value{
[matrix()] Time series as columns.
}
\description{
Simulate Three Species Dynamics via Hastings & Powell Model
}
\examples{
length_out <- 5000
time_step <- 1
burn_steps <- 5000
x0 <- 0.78
y0 <- 0.16
z0 <- 10.0
a1 <- 5.0
a2 <- 0.1
b1 <- 3.0
b2 <- 2.0
d1 <- 0.4
d2 <- 0.01

s1 <- pbsSim(
  length_out = length_out,
  time_step = time_step,
  burn_steps = burn_steps,
  x0 = x0,
  y0 = y0,
  z0 = z0,
  a1 = a1,
  a2 = a2,
  b1 = b1,
  b2 = b2,
  d1 = d1,
  d2 = d2)

}
