% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pbsSimplex}
\alias{pbsSimplex}
\title{Forecast via Simplex Projection}
\usage{
pbsSimplex(
  N,
  E = 1:10,
  p = 1L,
  first_difference = FALSE,
  centre_and_scale = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{N}{[data.frame()] or [numeric()] Response variable time series as
first column or as a vector.}

\item{E}{[integer()] Vector of embedding dimensions.}

\item{p}{[integer()] Scalar forecast distance.}

\item{first_difference}{[logical()] First-difference the time series}

\item{centre_and_scale}{[logical()] Centre and scale the time series?}

\item{verbose}{[logical()] Print progress?}
}
\value{
A list of class \code{pbsSimplex}
}
\description{
Forecast via Simplex Projection
}
\examples{
N <- data.frame(x = 1:30)
s1 <- pbsSimplex(N)

N <- data.frame(x = simple_ts)
s2 <- pbsSimplex(N)   

}
\author{
Luke A. Rogers
}
