% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{pbsSSR}
\alias{pbsSSR}
\title{Create State Space Reconstruction Matrix}
\usage{
pbsSSR(N, lags, p = 1L, first_difference = FALSE, centre_and_scale = FALSE)
}
\arguments{
\item{N}{[matrix()] or [data.frame()] with named [numeric()] columns 
for the response variable and covariate time series.}

\item{lags}{[list()] of named integer vectors specifying the lags to use for
each time series in \code{N}.}

\item{p}{[integer()] The forecast distance.}

\item{first_difference}{[logical()] First-difference each time series?}

\item{centre_and_scale}{[logical()] Centre and scale each time series?}
}
\value{
[matrix()] State space reconstruction.
}
\description{
Create State Space Reconstruction Matrix
}
