% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{pbsX}
\alias{pbsX}
\title{Create X Matrix}
\usage{
pbsX(Y, lags)
}
\arguments{
\item{Y}{[matrix()] with [numeric()] columns.}

\item{lags}{[list()] of named integer vectors specifying the lags to use for
each time series in \code{N}.}
}
\value{
[matrix()] X
}
\description{
Create X Matrix
}
