% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{pbsY}
\alias{pbsY}
\title{Create Y Matrix}
\usage{
pbsY(Z, centre_and_scale)
}
\arguments{
\item{Z}{[matrix()] with [numeric()] columns.}

\item{centre_and_scale}{[logical()] Centre and scale columns of Z?}
}
\value{
[matrix()] Y
}
\description{
Create Y Matrix
}
