% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{pct_error}
\alias{pct_error}
\title{Percent Error}
\usage{
pct_error(x, y)
}
\arguments{
\item{x}{[numeric()][vector()] observed values}

\item{y}{[numeric()][vector()] forecast values}
}
\value{
[numeric()][vector()] percent error
}
\description{
Percent Error
}
\examples{
pct_error(0:10, 10:0)

}
