% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.pbsEDM}
\alias{plot.pbsEDM}
\title{Plot data and results of an object of class `pbsEDM`}
\usage{
\method{plot}{pbsEDM}(obj, late.num = 5)
}
\arguments{
\item{obj}{list object of class `pbsEDM`, an output from `pbsEDM()`}
}
\value{

}
\description{
Only working if nt_observed values are in there
}
\author{
Andrew Edwards
}
