% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.pbsSim}
\alias{plot.pbsSim}
\title{Plot Method for `pbsSim`}
\usage{
\method{plot}{pbsSim}(x, ...)
}
\arguments{
\item{x}{[pbsSim()]}

\item{...}{Other arguments.}
}
\value{
Panel plot.
}
\description{
Plot Method for `pbsSim`
}
