% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.rho.Evec}
\alias{plot.rho.Evec}
\title{Plot rho versus E for output from pbsEDM_Evec
 <description>}
\usage{
\method{plot}{rho.Evec}(E_res, ...)
}
\arguments{
\item{E_res}{List of `pbsEDM` objects as output from `pbsEDM_Evec()`}
}
\value{
plot of rho versus E
}
\description{
Plot rho versus E for output from pbsEDM_Evec
 <description>
}
\examples{
\donttest{
  aa <- pbsEDM_Evec(Nx_lags_orig$Nt)
  plot.rho.Evec(aa)
}
}
\author{
Andrew Edwards
}
