% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_explain_edm}
\alias{plot_explain_edm}
\title{2-d phase plot of x(t) v x(t-1) with coloured points to explain EDM}
\usage{
plot_explain_edm(
  obj,
  tstar,
  x.lab = expression("X"[t - 1]),
  y.lab = expression("X"[t]),
  main = "All the points in lagged space",
  tstar.col = "blue",
  tstar.pch = 1,
  tstar.cex = 1,
  psivec = NULL,
  neigh.plot = FALSE,
  neigh.proj = FALSE,
  pred.plot = FALSE,
  pred.rEDM = FALSE,
  true.val = FALSE,
  legend.plot = TRUE
)
}
\arguments{
\item{obj}{list object of class `pbsEDM`, an output from `pbsEDM()`}

\item{tstar}{the time index (x(t)) of the target point for which to make a
projection from}
}
\value{

}
\description{
Highlights a point to be projected, its nearest `E+1` neighbours, and then
draw arrows to show where they go and so where the projection goes. Very
useful for understanding and checking what EDM is doing. Call this multiple
times using TODO to make a movie. Type `plot_explain_edm_movie` for example calls.
}
\author{
Andrew Edwards
}
