% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_explain_edm_all_tstar_movie_save}
\alias{plot_explain_edm_all_tstar_movie_save}
\title{Make pdf movie of explain EDM figure for each tstar to go into Supp pdf.}
\usage{
plot_explain_edm_all_tstar_movie_save(
  E_res,
  pdf.filename = "explain_EDM_all_tstar_movie.pdf",
  ...
)
}
\arguments{
\item{E_res}{List of `pbsEDM` objects as output from `pbsEDM_Evec()`}

\item{pdf.filename}{filename to save to}

\item{...}{extra arguments to pass to `plot_explain_edm_movie()`}
}
\value{
saved pdf file with one page for each figure, that can become
 movie in Supp of primer manuscript.
}
\description{
Creates a single .pdf with one frame for each figure (value of tstar),
 included as a pausable movie of Supp pdf for primer manuscript.
}
\examples{
\donttest{
  E_results <- pbsEDM_Evec(NY_lags_example$N_t)
  plot_explain_edm_all_tstar_movie_save(E_results[[1]])
}
}
\author{
Andrew Edwards
}
