% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_explain_edm_movie}
\alias{plot_explain_edm_movie}
\title{Create movie to explain EDM}
\usage{
plot_explain_edm_movie(obj, ...)
}
\arguments{
\item{obj}{list object of class `pbsEDM`, an output from `pbsEDM()`}

\item{...}{needs to include tstar}
}
\value{

}
\description{
Use with gifski in .Rmd - see vignette.
}
\author{
Andrew Edwards
\donttest{
  aa <- pbsEDM_Evec(Nx_lags_orig$Nt)
  plot_explain_edm_movie(aa, tstar = 15)
}
}
