% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_explain_edm_movie_save}
\alias{plot_explain_edm_movie_save}
\title{Make pdf movie of explain EDM figures to go into Supp pdf.}
\usage{
plot_explain_edm_movie_save(
  E_res,
  pdf.filename = "explain_EDM_movie.pdf",
  tstar = 39,
  ...
)
}
\arguments{
\item{E_res}{List of `pbsEDM` objects as output from `pbsEDM_Evec()`}

\item{pdf.filename}{filename to save to}

\item{tstar}{focal point to make projection from}

\item{...}{extra arguments to pass to `plot_explain_edm_movie()`}
}
\value{
saved pdf file with one page for each figure, that can become
 movie in Supp of primer manuscript.
}
\description{
Creates a single .pdf with one frame for each figure,
 included as a pausable movie of Supp pdf for primer manuscript.
 See `analyse_simple_time_series.Rmd` vignette for creating the similar .gif
 that is shown in the vignette.
}
\examples{
\donttest{
  E_results <- pbsEDM_Evec(NY_lags_example$N_t)
  plot_explain_edm_movie_save(E_results[[1]])
}
}
\author{
Andrew Edwards
}
