% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_explain_edm_save}
\alias{plot_explain_edm_save}
\title{Create and save encapsulated postscript file for manuscript to explain EDM.}
\usage{
plot_explain_edm_save(E_res, eps.filename = "explain_EDM.eps", tstar = 39)
}
\arguments{
\item{E_res}{List of `pbsEDM` objects as output from `pbsEDM_Evec()`}

\item{eps.filename}{filename to save to}

\item{tstar}{focal point to make projection from}
}
\value{
saved .eps file to use in manuscript.
}
\description{
Create and save encapsulated postscript file for manuscript to explain EDM.
}
\examples{
\donttest{
# For manuscript:
E_results <- pbsEDM_Evec(NY_lags_example$N_t)
plot_explain_edm_save(E_results[[1]])
}
}
\author{
Andrew Edwards
}
