% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_library_size}
\alias{plot_library_size}
\title{Plot the size of the library for a given `T` as a function of `E` and `tstar`}
\usage{
plot_library_size(
  T = 50,
  E_vec = 2:10,
  tstar_vec = 1:50,
  annotate = TRUE,
  annotate_cex = 1,
  annotate_tstar = 23,
  annotate_extra_cex = 0.7,
  words_pos = c(6, 3.5, 6, 49.6),
  words_cex = 0.9,
  mgp_vals = c(2, 0.5, 0)
)
}
\arguments{
\item{T}{integer of the sample size of a univariate time series (time series
itself not needed)}

\item{E_vec}{vector of integers to show for the embedding dimension}

\item{tstar_vec}{vector of integers to show for the focal point time index
`tstar`, `t*` in the manuscript, from which projections are made from}

\item{annotate}{logical whether to add numbers along the middle of the plot
(avoids the need for a colorbar, which was fiddly to do, see attempts
deleted in 8f567ff, while also having a grey background), and add text to
explain the grey areas.}

\item{annotate_cex}{text size for main annotation}

\item{annotate_tstar}{tstar value at which to add the annotated numbers; if
NULL then is `max(tstar_vec)/2`.}

\item{annotate_extra_cex}{text size for extra smaller numbers in boxes at
the top}

\item{words_pos}{vector of positioning for words in grey areas, given by `E`
and `tstar` position for `Early values' text (text is placed to the right
of these), then for `Late values' text (text is centred around these.}

\item{words_cex}{text size for above words}

\item{mgp_vals}{mgp values to use, as in `plot(..., mgp = mgp_vals)`.}
}
\value{
plots the contour plot and returns the matrix (`C` in manuscript) of calculated library sizes
}
\description{
Show a coloured contour plot of how the library size changes for a
univariate time series with sample size `T` as chosen embedding dimension
`E` and target index `tstar` vary.
}
\examples{
\dontrun{
# For manusript doing:
postscript("library_size.eps",
            height = 6,
            width = 6,
            horizontal=FALSE,
            paper="special")
plot_library_size()
dev.off()
}
}
\author{
Andrew Edwards
}
