% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_library_size_save}
\alias{plot_library_size_save}
\title{Save eps figure of the size of the library for manuscript.}
\usage{
plot_library_size_save(eps.filename = "library_size.eps")
}
\arguments{
\item{eps.filename}{filename to save to}
}
\value{
saved .eps file to use in manuscript.
}
\description{
Save eps figure of the size of the library for manuscript.
}
\author{
Andrew Edwards
}
