% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-mve-spaces.R
\name{plot_mve_spaces}
\alias{plot_mve_spaces}
\title{Plot the axes used in the top multiview embedding reconstructions}
\usage{
plot_mve_spaces(
  res_mve,
  col = "blue",
  rho_on_top = TRUE,
  max_embeddings = 50,
  first_embedding = 1,
  ...
)
}
\arguments{
\item{res_mve}{list object resulting from a call to [multiview_embedding()]}

\item{col}{character; color to use for tiles where the lag is `TRUE`. Default "blue".}

\item{rho_on_top}{logical; if `TRUE` (default), rho values are along the top (x-axis).
If `FALSE`, rho values are on the y-axis.}

\item{max_embeddings}{maximum number of reconstructions to show, so as to not make
the plot massive.}

\item{first_embedding}{the first embedding to use for the plot. So if you
want to show 100 across two plots, use `first_embedding = 51` for the second plot.}

\item{...}{passed onto [create_named_lags()], namely `order` to order the
variables consistently.}
}
\value{
ggplot object
}
\description{
Takes the results list from [multiview_embedding()] and indicates the
variables and lags used in the top reconstruction subsets.
}
\examples{
\dontrun{
  plot_mve_spaces(age11_res)
  plot_mve_spaces(age11_res, rho_on_top = FALSE)
}

}
\author{
Andrew M. Edwards and GitHub Copilot
}
