% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_pbsEDM}
\alias{plot_pbsEDM}
\title{<description>}
\usage{
plot_pbsEDM(obj)
}
\arguments{
\item{obj}{list object of class `pbsEDM`, an output from `pbsEDM()`}
}
\value{

}
\description{
<description>
}
\author{
Andrew Edwards
}
