% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_pbsEDM_Evec_movie}
\alias{plot_pbsEDM_Evec_movie}
\title{Make pdf movie of six-panel figure to go into Supp pdf.}
\usage{
plot_pbsEDM_Evec_movie(E_res, pdf.filename = "six_panel_movie.pdf", ...)
}
\arguments{
\item{E_res}{List of `pbsEDM` objects as output from `pbsEDM_Evec()`}

\item{pdf.filename}{filename to save to}

\item{...}{extra arguments to pass to `plot.pbsEDM()`}
}
\value{
saved pdf file with one page for each time step, that can become
 movie in Supp of primer manuscript.
}
\description{
Creates a single .pdf with one frame for each time step, which can be
 included as a pausable movie of Supp pdf for primer manuscript.
 See `analyse_simple_time_series.Rmd` vignette for creating the similar .gif
 that is shown in the vignette.
}
\examples{
\donttest{
  E_results <- pbsEDM_Evec(NY_lags_example$N_t)
  plot_pbsEDM_Evec_movie(E_results)
}
}
\author{
Andrew Edwards
}
