% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_pbsEDM_Evec_save}
\alias{plot_pbsEDM_Evec_save}
\title{Plot output from `plot_pbsEDM_Evec()` as six panel plot saved as
 encapsulated postscript file for manuscript.}
\usage{
plot_pbsEDM_Evec_save(E_res, eps.filename = "six_panels_all.eps", ...)
}
\arguments{
\item{E_res}{List of `pbsEDM` objects as output from `pbsEDM_Evec()`}

\item{eps.filename}{filename to save to}

\item{...}{extra arguments to `plot_pbsEDM_Evec()`.}
}
\value{
Saved .eps file to use in manuscript
}
\description{
Plot output from `plot_pbsEDM_Evec()` as six panel plot saved as
 encapsulated postscript file for manuscript.
}
\examples{
\donttest{
E_results <- pbsEDM_Evec(NY_lags_example$N_t)
plot_pbsEDM_Evec_save(E_results)
}
}
\author{
Andrew Edwards
}
