% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_pred_obs}
\alias{plot_pred_obs}
\title{Plot predictions versus observed values of `Y_t` for list of `pbsEDM`
objects, for various values of `E`}
\usage{
plot_pred_obs(
  E_res,
  E_components = 5,
  E_cols = c("orange", "blue", "green", "red", "black"),
  last.time.to.plot = NULL,
  E_cex = seq(1.7, 0.6, length = 5),
  portrait = NULL,
  label = NULL,
  label.cex = 0.7
)
}
\arguments{
\item{E_res}{List of `pbsEDM` objects as output from `pbsEDM_Evec()`}

\item{E_components}{How many of the first E_res components to show}

\item{E_cols}{Vector of colours, one for each value of E}

\item{last.time.to.plot}{Last time value of `t` to use when plotting.}

\item{E_cex}{Vector of sizes, one for each value of E}

\item{portrait}{dummy argument that allows `...` to be passed from
`plot_pbsEDM_Evec()`, from which we want `last.time.to.plot`.}

\item{label}{label to annotate plot, such as `(a)` etc. for six-panel figure}

\item{label.cex}{size of label annotation}
}
\value{
Figure in the current plot enivironment
}
\description{
Plot the predicted versus observed values of `Y_t` for various values
of `E`, using the output (a list of `pbsEDM` objects) from
`pbsEDM_Evec()`. Each value of `t` corresponds to focal point `t* = t - 1`.
}
\examples{
\donttest{
  aa <- pbsEDM_Evec(NY_lags_example$N_t)
  plot_pred_obs(aa)
}
}
\author{
Andrew Edwards
}
