% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_rho_Evec}
\alias{plot_rho_Evec}
\title{Plot rho versus `E` for output from `pbsEDM_Evec()`}
\usage{
plot_rho_Evec(E_res, ...)
}
\arguments{
\item{E_res}{List of `pbsEDM` objects as output from `pbsEDM_Evec()`}

\item{...}{Extra arguments to pass to `plot()`}
}
\value{
plot of rho versus E
}
\description{
Plot correlation coefficient against `E`.
}
\examples{
\donttest{
E_results_17 <- pbsEDM_Evec(NY_lags_example$N_t,
                            E_vec = 1:16)  # go up to E=17
  postscript("rho_Evec.eps",
            height = 4.5,
            width = 4.5,
            horizontal=FALSE,
            paper="special")
  plot_rho_Evec(E_results_17)
  dev.off()
}
}
\author{
Andrew Edwards
}
