% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_rho_Evec_save}
\alias{plot_rho_Evec_save}
\title{Save `rho` versus `E` .eps figure for manuscript.}
\usage{
plot_rho_Evec_save(eps.filename = "rho_Evec.eps")
}
\arguments{
\item{eps.filename}{filename to save to}
}
\value{
plot of rho versus E saved as .eps
}
\description{
Save `rho` versus `E` .eps figure for manuscript.
}
\examples{
\donttest{
}
}
\author{
Andrew Edwards
}
