% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotPanelMovie.df2}
\alias{plotPanelMovie.df2}
\title{Plot a movie of data (and EDM results if available) with panels showing data
 in various ways}
\usage{
plotPanelMovie.df2(Nx.lags = Nx_lags_orig,
  pdf.filename = "Nx_lags_orig_movie.pdf", rhoForE = NULL,
  Evec = NULL, Ecols = NA, includeTimeSeries = TRUE,
  final_plot = "compare", start = 1, end = nrow(Nx.lags),
  max_time = NULL, only.final.plot = FALSE, cobwebbing = TRUE,
  late.col = "red", early.col = "black",
  early.col.lines = "lightgrey", late.num = 3, pt.type = "p",
  x.lab = expression("x"[t - 2]), y.lab = expression("x"[t - 1]),
  z.lab = expression("x"[t]), figheight = 9.8, figwidth = 6.84,
  open.pdf = TRUE, ...)
}
\arguments{
\item{Nx.lags}{Dataframe (Tibble???) with time indexing the row, and columns
  named Nt, Ntmin1, Xt, Xtmin1, Xtmin2 and, if E!=0, columns XtPredEeq1,
  XtPredEeq2, etc. TODO - not same now, using saved example but also want
option for doing Eeq2 etc. from rEDM and from pbsEDM. TODO TODO
  This allows Xt to equal Nt if want to plot figures
  without the first differencing (basically the 3d plot and the pred
  vs obs would be the only new ones). Nt is the original data,
  Xt is the first difference, and Xt.est is the estimate for Xt from
  EDM. Row numbers are time, and start at 1.
  Could maybe extend to have multiple Xt.est for the different
  embedding dimensions. TODO}

\item{pdf.filename}{filename (including .pdf) to save the movie as if saving
as .pdf. Ignored if open.pdf is FALSE.}

\item{rhoForE}{values of rho corrsponding to each E in Evec}

\item{Evec}{vector of E values used to construct columns in Nx.lags called
XtPredEeq1, XtPredEeq2 etc. (Xt predicted with E=1, etc.). If NULL
then no predicted vs observed plot is drawn.}

\item{Ecols}{colour coding for E points in predicted vs observed plot}

\item{includeTimeSeries}{TRUE to include the time series at the top (FALSE
hasn't  been tested yet TODO)}

\item{final_plot}{What the final plot should be. Currently only works for
`compare` to compare Andy's predictions with rEDM's. Will then extend to
compare different E values specified in `Evec` with `final_plot` being "E_vary".}

\item{start}{first time value (row) to use when plotting - not fully
implemented as the colours aren't correct. Gives warning if != 1. TODO}

\item{end}{last time value (row) to use when plotting (so will have
end-start points plotted) TODO (end-start-1??).}

\item{max_time}{maximum time value to plot (will be different from `end` if
making a movie using Rmarkdown, since for that `end` will get iterated)}

\item{only.final.plot}{if TRUE then only plot the final plot}

\item{cobwebbing}{whether to join up points in 2d lag plots via cobwebbing;
if FALSE then joins consecutive points (TODO?? - should be TRUE?).}

\item{late.col}{colour for the late points (the last 'late.num' points)}

\item{early.col}{colour for the early points}

\item{early.col.lines}{colour for the early lines}

\item{late.num}{how many points to have colour 'late.col'}

\item{pt.type}{usual 'type' option (but don't use "o")}

\item{x.lab}{x-label for axes for 3d lagged plots - for 2d
lagged plots z.lab is the y label and y.lab is the x label.}

\item{y.lab}{y-label for axes for 3d lagged plots - for 2d
lagged plots z.lab is the y label and y.lab is the x label.}

\item{z.lab}{y-label for axes for 3d lagged plots - for 2d
lagged plots z.lab is the y label and y.lab is the x label.}

\item{figheight}{height of .pdf, but ignored if open.pdf == FALSE.}

\item{figwidth}{width of .pdf, but ignored if open.pdf == FALSE.}

\item{open.pdf}{whether or not to open a new .pdf file for the figure. Use
FALSE if .pdf is being opened automatically in knitr (then figheight
and figwidth get ignored, and should be set in <<..>>=.}
}
\value{
Creates movie as a .pdf then can then be easily input into Sweave file
    using the animategraphics latex pacakge. TODO make .gif option also for vignette.
}
\description{
Create figure with six panels as a movie. Panels are:
N_t vs time and x_t vs time (optional)
N_t vs N_{t-1}, x_t vs x_{t-1} (like plotLag2d),
x_t vs x_{t-1} vs x_{t-2} (like plotLag3d) and
pred vs obs (for given values of E, if provided).
All code is in here, may want to parse out into individual plots to be able
to more easily reuse.
With colours to show the latest points, and star to highlight latest point.
Defaults are for `Nx_lags_orig` example.
}
\details{
TODO: change options ot have NULl if depend on earlier things. end may need
to be one less for default.
TODO: change . to _ everywhere
TODO: options to save as pdf (maybe - used to be for Sweave but now
superceded) or gif to be in a vignette.
TODO: adapt to work with Luke's output formats, or just write a wrapper
function. Most of this is just looking at data in various ways.
}
\author{
Andrew Edwards
}
