% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{raw_error}
\alias{raw_error}
\title{Raw Error}
\usage{
raw_error(x, y)
}
\arguments{
\item{x}{[numeric()][vector()] observed values}

\item{y}{[numeric()][vector()] forecast values}
}
\value{
[numeric()][vector()] raw error
}
\description{
Raw Error
}
