% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rEDM_points}
\alias{rEDM_points}
\title{Predicted and Observed Values from rEDM output}
\format{
Dataframe of size 100x4 with one row for each `time`, and columns:
  * `time`
  * `obs`
  * `pred`
  * `pred_var`
}
\usage{
rEDM_points
}
\description{
Predicted and Observed Values from rEDM output
}
\keyword{datasets}
