% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{simplex_forecast}
\alias{simplex_forecast}
\title{Return Simplex Projection Forecasts from Nearest Neighbour Values and Weights}
\usage{
simplex_forecast(value_matrix, weight_matrix)
}
\arguments{
\item{value_matrix}{A matrix of row vectors giving the projected nearest 
neighbour values}

\item{weight_matrix}{A matrix of row vectors giving the projected nearest 
neighbour weights}
}
\value{
A vector of forecasts for the projected values
}
\description{
Return Simplex Projection Forecasts from Nearest Neighbour Values and Weights
}
\examples{
A <- matrix(c(12, 15, 13,
              21, 19, 22,
              28, 29, 25), nrow = 3, byrow = TRUE)
W <- matrix(c(0.6, 0.3, 0.1,
              0.5, 0.4, 0.1,
              0.7, 0.2, 0.1), nrow = 3, byrow = TRUE)
simplex_forecast(A, W)

}
