% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mve.R
\name{single_view_embedding}
\alias{single_view_embedding}
\title{Single View Embedding}
\usage{
single_view_embedding(
  data,
  response,
  lags,
  index = 50L,
  buffer = 10L,
  window = integer(0),
  metric = "mamse",
  beyond = FALSE,
  superset = NULL
)
}
\arguments{
\item{data}{[matrix()] or [data.frame()] with named [numeric()] columns}

\item{response}{[character()] column name of the response variable in
\code{data}}

\item{lags}{[list()] of a named vector of lags for each explanatory
variable.}

\item{index}{[integer()]}

\item{buffer}{[integer()] number of forecasts prior to \code{index}}

\item{window}{[integer()] forecast metric moving window width}

\item{metric}{[character()]}

\item{beyond}{[logical()]}

\item{superset}{[list()] superset of lags corresponding to parent SSR}
}
\value{
[data.frame()]
}
\description{
Single View Embedding
}
\author{
Luke A. Rogers
}
