% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{smap_efficient}
\alias{smap_efficient}
\title{SMap That Returns Only The Forecast Accuracy}
\usage{
smap_efficient(
  N,
  lags = NULL,
  theta = 0,
  p = 1L,
  first_difference = FALSE,
  centre_and_scale = FALSE
)
}
\arguments{
\item{N}{A data frame with named columns for the response variable and
covariate time series.}

\item{lags}{A list of named integer vectors specifying the lags to use for
each time series in \code{N}.}

\item{theta}{The numeric local weighting parameter.}

\item{p}{The integer forecast distance.}

\item{first_difference}{Logical. First-difference each time series?}

\item{centre_and_scale}{Logical. Centre and scale each time series?}
}
\value{
[numeric()] The forecast accuracy rho
}
\description{
SMap That Returns Only The Forecast Accuracy
}
\examples{
\donttest{
 N <- data.frame(x = simple_ts)
 lags <- list(x = 0:1)
 m4 <- smap_efficient(N, lags, first_difference = TRUE)
}
# And see pbsSmap vignette.
}
