% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_previous.R
\name{smap_forecast}
\alias{smap_forecast}
\title{Return S-Map Forecasts from SVD Solutions and Embedded Points}
\usage{
smap_forecast(c_matrix, focal_matrix)
}
\arguments{
\item{c_matrix}{A matrix of row vector SVD solutions}

\item{focal_matrix}{A matrix of row vectors giving the focal point from 
which to forecast}
}
\value{

}
\description{
Return S-Map Forecasts from SVD Solutions and Embedded Points
}
\examples{
c_mat <- matrix(c(0.4, 0.3, 0.3,
                  0.5, 0.4, 0.1,
                  0.7, 0.2, 0.1), nrow = 3, byrow = TRUE)
f_mat <- matrix(c(16, 11, 19,
                  21, 14, 26,
                  28, 26, 27), nrow = 3, byrow = TRUE)
smap_forecast(c_mat, f_mat)
                  
}
