% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smap_thetavec.R
\name{smap_thetavec}
\alias{smap_thetavec}
\title{Do the S-map calculations for a vector of `theta` values}
\usage{
smap_thetavec(N, lags, theta_vec = seq(0, 1, by = 0.1), ...)
}
\arguments{
\item{N}{A data frame with named columns for the response variable and
covariate time series.}

\item{lags}{A list of named integer vectors specifying the lags to use for
each time series in \code{N}.}

\item{theta_vec}{Vector of theta values to use.}

\item{...}{Further options to pass to `smap_efficient()`. In particular
`first_difference` may need to be TRUE, the default is FALSE.}
}
\value{
Vector of values of `rho` corresponding to each value of `theta_vec`.
}
\description{
Given a vector of theta values and data, do the S-map calculation for each theta, using the
fast `smap_efficient()`. If values are not already first-differenced then set
`first_difference = TRUE` so the differencing is done within
`smap_efficient()`.
}
\examples{
\donttest{
 N <- data.frame(x = simple_ts)
 lags <- list(x = 0:1)
 res <- smap_thetavec(N, lags, theta_vec = seq(0, 2, by=0.1), first_difference = TRUE)
 plot(res)
}
# And see pbsSmap vignette.
}
\author{
Andrew Edwards
}
