% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{srmse}
\alias{srmse}
\title{Standardised Root Mean Squared Error}
\usage{
srmse(x, y)
}
\arguments{
\item{x}{[numeric()][vector()] observed values}

\item{y}{[numeric()][vector()] forecast values}
}
\value{
[numeric()] root mean squared error
}
\description{
Standardised Root Mean Squared Error
}
\examples{
x <- rnorm(100, 0, 1)
y1 <- x
srmse(x, y1)

y2 <- rep(mean(x), length(x))
srmse(x, y2)

}
