% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_space_distances.R
\name{state_space_distances}
\alias{state_space_distances}
\title{State Space Distance Matrix}
\usage{
state_space_distances(ssr, index = 50L, buffer = 10L)
}
\arguments{
\item{ssr}{[matrix()] a state space reconstruction in which the rows
are points in the state space}

\item{index}{[integer()] time index of the first value to forecast}

\item{buffer}{[integer()] number of values to forecast before \code{index}}
}
\value{
[matrix()] of allowed neighbour distances, rows corresponding to
  focal point time, columns to neighbour point time.
}
\description{
State Space Distance Matrix
}
\details{
Row index corresponds to focal point time. Column index
  corresponds to neighbour point time. The value represents the distance
  from the focal point to the neighbour point. Disallowed focal point
  and neighbour combinations have value NA.
}
\author{
Luke A. Rogers
}
