% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_space_forecasts.R
\name{state_space_forecasts}
\alias{state_space_forecasts}
\title{Empirical Dynamic Modeling Forecasts}
\usage{
state_space_forecasts(X, distance, beyond = FALSE)
}
\arguments{
\item{X}{[matrix()] a state space reconstruction in which the rows
are points in the state space}

\item{distance}{[matrix()] of allowed neighbour distances}

\item{beyond}{[logical()]}
}
\value{
[numeric()] [vector()] of forecast values
}
\description{
Empirical Dynamic Modeling Forecasts
}
\author{
Luke A. Rogers
}
