% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{state_space_reconstruction}
\alias{state_space_reconstruction}
\title{State Space Reconstruction (SSR)}
\usage{
state_space_reconstruction(data, response, lags)
}
\arguments{
\item{data}{[matrix()] with variables as named columns}

\item{response}{[character()] column name of the response variable}

\item{lags}{[list()] of a named vector of lags for each explanatory variable}
}
\value{
[state_space_reconstruction()] [matrix()] with unlagged response
  and lagged explanatory variables centred on their means and scaled by
  their respective standard deviations, with automatically generated column
  names.
}
\description{
Rows are centred and scaled points in the state-space
  reconstruction.
}
\examples{
d <- data.frame(x = 1:10, y = 11:20)
state_space_reconstruction(d, response = "x", lags = list(y = c(0, 1, 2, 3)))

}
\author{
Luke A. Rogers
}
