% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{superset_columns}
\alias{superset_columns}
\title{Lag Superset Columns}
\usage{
superset_columns(data, lags, superset = NULL, beyond = FALSE)
}
\arguments{
\item{data}{[matrix()] or [data.frame()] with named [numeric()] columns}

\item{lags}{[list()] of a named vector of lags for each explanatory
variable.}

\item{superset}{[list()] superset of lags corresponding to the parent state
space reconstruction}

\item{beyond}{[logical()]}
}
\value{
[tibble::tibble()]
}
\description{
Lag Superset Columns
}
