% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{untransform_forecasts}
\alias{untransform_forecasts}
\title{Untransform State Space Forecast Values}
\usage{
untransform_forecasts(x, y)
}
\arguments{
\item{x}{[numeric()][vector()] observed values}

\item{y}{[numeric()][vector()] state space forecast values}
}
\value{
[numeric()][vector()]
}
\description{
Untransform State Space Forecast Values
}
