% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_previous.R
\name{util_exclude_indices}
\alias{util_exclude_indices}
\title{Exclude Indicies Associated with a Specified Time Index}
\usage{
util_exclude_indices(time_ind, pred_dist, lag_size, embed_dim, one_sided = T)
}
\arguments{
\item{time_ind}{The time index of the focal vector (numeric scalar)}

\item{pred_dist}{The prediction distance (numeric scalar)}

\item{lag_size}{The number of time steps separating successive lags}

\item{embed_dim}{The embedding dimension (numeric scalar)}

\item{one_sided}{Should only the focal value be excluded? (logical)}
}
\value{
A numeric vector
}
\description{
Exclude Indicies Associated with a Specified Time Index
}
\examples{
util_exclude_indices(10, 1, 1, 5, T)

}
