% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_single_view_embeddings.R
\name{weight_single_view_embeddings}
\alias{weight_single_view_embeddings}
\title{Weight Single View Embeddings}
\usage{
weight_single_view_embeddings(forecasts, metric, weight, n_weight)
}
\arguments{
\item{forecasts}{[list()]}

\item{metric}{[character()]}

\item{weight}{[character()]}

\item{n_weight}{[numeric()]}
}
\value{
[list()]
}
\description{
Weight Single View Embeddings
}
