% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_subset_lags.R
\name{create_subset_lags}
\alias{create_subset_lags}
\title{Create Subset Lags}
\usage{
create_subset_lags(lags, response_name = NULL)
}
\arguments{
\item{lags}{[list()] of a named vector of lags for each explanatory
variable.}

\item{response}{character string; if not NULL then any subset that does not
include `response` with a lag of 0 is excluded; this if for running
`pbsEDM` (which requires response variable in the state space) in
`multivariate_simplex()`.}
}
\value{
[list()] with one lag [list()] for each subset state space
  reconstruction
}
\description{
Creates all combinations, though not in quite the logical order one might
  expect. Could resolve by reording the code maybe, but it works (Andy
  manually checked a detailed example, essentially the one given).
}
\examples{
s <- create_subset_lags(list(a = c(0, 1, 2), b = c(0, 1)))
s_with_response <- create_subset_lags(list(a = c(0, 1, 2), b = c(0, 1)), response = "b")
}
