% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_metrics.R
\name{forecast_metrics}
\alias{forecast_metrics}
\title{Forecast Metrics With Rolling Windows (we don't need for salmon)}
\usage{
forecast_metrics(x, y, k = integer(0), metric = "mamse")
}
\arguments{
\item{x}{[numeric()] [vector()] observed values}

\item{y}{[numeric()] [vector()] forecast values}

\item{k}{[numeric()] scalar size of the trailing running window. Uses all
previous data when \code{k = integer(0)}.  TODO we don't want this.}

\item{metric}{[character()]}
}
\value{
[tibble::tibble()]
}
\description{
Forecast Metrics With Rolling Windows (we don't need for salmon)
}
