% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_previous.R
\name{make_dist_tibble}
\alias{make_dist_tibble}
\title{Make a Tibble of Euclidean Distances Between Rows of a Tibble}
\usage{
make_dist_tibble(tbl)
}
\arguments{
\item{tbl}{A tibble of row vectors (tibble)}
}
\value{
A tibble of Euclidean distances
}
\description{
Make a Tibble of Euclidean Distances Between Rows of a Tibble
}
\examples{
  tbl <- tibble::tibble(x = 1:6, y = c(NA, 8:12), z = 13:18)
  dist_tibble <- make_dist_tibble(tbl)

}
