% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{make_global_indices}
\alias{make_global_indices}
\title{Make Global Allowable Indices from a Matrix of Lagged Column Row Vectors}
\usage{
make_global_indices(mat, from = seq_len(nrow(mat)),
  into = seq_len(nrow(mat)), dist = 1L)
}
\arguments{
\item{mat}{A matrix of lagged columns defining row vectors (matrix)}

\item{from}{A vector of indices to consider forecasting from (numeric)}

\item{into}{A vector of indices to consider forecasting into (numeric)}

\item{dist}{The forecast distance (numeric scalar)}
}
\value{
A tibble with two columns of indices
}
\description{
Make Global Allowable Indices from a Matrix of Lagged Column Row Vectors
}
\examples{
dat <- data.frame(x = c(1:7, NA, 9:15), y = 11:25)
mat <- as.matrix(combine_lag_tibbles(dat, c("x", "y"), c(3, 2), 1))
make_global_indices(mat)

}
