% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{make_lag_tibble}
\alias{make_lag_tibble}
\title{Make a Tibble of Lagged Columns of a Data Frame}
\usage{
make_lag_tibble(data, lags)
}
\arguments{
\item{data}{A data frame with a named numeric columns to be lagged}

\item{lags}{A named list giving the lags to use for each corresponding
column. List names must match column names. (list of numeric vectors)}
}
\value{
A tibble of lagged columns
}
\description{
Make a Tibble of Lagged Columns of a Data Frame
}
\examples{
dat <- data.frame(x = 1:10, y = 11:20, z = 21:30)
make_lag_tibble(dat, list(x = 0:2, y = 0:1, z = 0))

}
