% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{make_local_indices}
\alias{make_local_indices}
\title{Make Local Allowable Indices for Nearest Neighbour Vectors}
\usage{
make_local_indices(index, from_global, lags, dist = 1L, symm = FALSE)
}
\arguments{
\item{index}{Index for the focal vector (numeric scalar)}

\item{from_global}{A vector of indices to consider forecasting from (numeric)}

\item{lags}{Named list of numeric vectors giving lags for each column. 
List names must match column names. (list of numeric vectors)}

\item{dist}{The forecast distance (numeric scalar)}

\item{symm}{Symmetric exclusion radius? (logical scalar)}
}
\value{
A vector of allowable indices for nearest neighbours
}
\description{
Make Local Allowable Indices for Nearest Neighbour Vectors
}
\examples{
make_local_indices(8, 1:15, list(x = 0:2))

}
