% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{make_simplex_forecast}
\alias{make_simplex_forecast}
\title{Title}
\usage{
make_simplex_forecast(from_index, from_global, lags, lag_tibble,
  distance_tibble, forecast_distance, symmetric_exclusion = FALSE)
}
\arguments{
\item{from_index}{}

\item{from_global}{}

\item{lags}{}

\item{lag_tibble}{}

\item{distance_tibble}{}

\item{forecast_distance}{}

\item{symmetric_exclusion}{}
}
\value{
A list
}
\description{
Title
}
\examples{

}
