% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbs_edm.R
\name{pbs_edm}
\alias{pbs_edm}
\title{Use EDM to Perform Out-of-Sample Forecasting}
\usage{
pbs_edm(data_frame, lags, from_user = seq_len(nrow(data_frame)),
  into_user = seq_len(nrow(data_frame)), forecast_distance = 1L,
  symmetric_exclusion = FALSE, include_stats = TRUE,
  include_forecasts = TRUE, include_neighbours = TRUE)
}
\arguments{
\item{data_frame}{A data frame with named columns (list of numeric vectors)}

\item{lags}{Named list of numeric vectors giving lags to use for each
column. List names must match column names. (list of numeric vectors)}

\item{from_user}{Rows to predict from (numeric vector)}

\item{into_user}{Rows to predict into (numeric vector)}

\item{forecast_distance}{Forecast distance (numeric scalar)}

\item{symmetric_exclusion}{Symmetric exclusion radius? (logical scalar)}

\item{include_stats}{Return forecast stats? (logical scalar)}

\item{include_forecasts}{Return forecast and observations? (logical scalar)}

\item{include_neighbours}{Return neighbours? (logical scalar)}
}
\value{
A tibble
}
\description{
Use EDM to Perform Out-of-Sample Forecasting
}
\examples{

}
