% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbs_edm.R
\name{pbs_edm}
\alias{pbs_edm}
\title{Use EDM to Perform Out-of-Sample Forecasting}
\usage{
pbs_edm(data, names, dims, lag = 1L, dist = 1L, symm = FALSE,
  from = seq_len(nrow(mat)), into = seq_len(nrow(mat)), stats = TRUE)
}
\arguments{
\item{data}{A data frame with named columns (list of numeric vectors)}

\item{names}{Names of numeric columns in data (character vector)}

\item{dims}{Number of embedding dimensions for each column (numeric vector)}

\item{lag}{Number of time steps separating successive lags (numeric scalar)}

\item{dist}{Forecast distance (numeric scalar)}

\item{symm}{Symmetric exclusion radius? (logical scalar)}

\item{from}{Rows to predict from (numeric vector)}

\item{into}{Rows to predict into (numeric vector)}

\item{stats}{Return only stats and omit neighbour lists? (logical scalar)}
}
\value{
A data frame
}
\description{
Use EDM to Perform Out-of-Sample Forecasting
}
\examples{

}
