% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbs_make_lags.R
\name{pbs_make_lags}
\alias{pbs_make_lags}
\title{Create a matrix of lags of an original time series vector}
\usage{
pbs_make_lags(time_series, embed_dim, lag_size)
}
\arguments{
\item{time_series}{A numeric vector with possible NA values}

\item{embed_dim}{An integer embedding dimension}

\item{lag_size}{The number of time steps separating successive lags}
}
\description{
Create a matrix of lags of an original time series vector
}
