% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{pbsEDM}
\alias{pbsEDM}
\title{Perform Out-of-Sample Forecasting via Empirical Dynamic Modelling}
\usage{
pbsEDM(
  nt,
  lags,
  forecast_distance = 1L,
  first_difference = FALSE,
  centre_and_scale = FALSE
)
}
\arguments{
\item{nt}{[data.frame()] A data frame with named columns for the raw (unlagged) variables}

\item{lags}{[list()] A named list of integer vectors specifying the lags for
each of the variables that will be used (other variables are ignored)}

\item{forecast_distance}{[integer(1)] The forecast distance}

\item{first_difference}{[logical(1)] First difference each variable before lagging?}

\item{centre_and_scale}{[logical(1)] Centre and scale each variable before lagging?}

\item{show_calculations}{[logical(1)]}
}
\value{
A list containing:

\itemize{
  \item nt_observed = NULL,
  \item nt_forecast = NULL,
  \item xt_observed [numeric()] Observed (possibly transformed) with \code{length = nrow(nt)}
  \item xt_forecast [numeric()] Forecast (of possibly transformed) with \code{length = nrow(nt)}
  \item xt_lags [matrix()] A column matrix of lagged variables with \code{nrow = nrow(nt)}
  \item xt_distance [matrix()] Focal index (row), neighbour (column),
  distance (value); large square matrix with \code{dim = nrow(nt)}
  \item xt_nbr_index [matrix()] Focal index (row), rank (column), and neighbour index (value)
  \item xt_nbr_value [matrix()] Focal index (row), rank (column), and
  neighbour value (value)  TODO Andy not quite sure what this is, don't you
  need X(t) and X(t-1) if, say, E=2?
  \item xt_nbr_distance [matrix()] Focal index (row), rank (column), and neighbour distance (value)
  \item xt_nbr_weight [matrix()] Focal index (row), rank (column), and neighbour weight (value)
  \item xt_prj_index [matrix()] Projected index (row), rank (column), and neighbour index (value)
  \item xt_prj_value [matrix()] Projected index (row), rank (column), and neighbour index (value)
  \item xt_prj_weight  [matrix()] Projected index (row), rank (column), and neighbour weight (value)
  \item forecast_distance = [integer(1)]
  \item first_difference = [logical(1)]
  \item centre_and_scale = [logical(1)]
  \item results [data.frame()] One-row data frame with values of E, rho and rmse.
}
}
\description{
Perform Out-of-Sample Forecasting via Empirical Dynamic Modelling
}
\details{
Only lags of variables explicitly named in \code{lags} are used. The
observed time series whose values are forecast must be specified first. For
example, given a data frame with variables \code{Predator}, \code{Prey} and \code{Salinity},
the unlagged version of the variable \code{Predator} can be specified as the
observed time series by

\code{lags = list(Predator = c(0, ...), ...)}.

The unlagged version of \code{Predator} can be forecast from the first and second lags
of \code{Predator}, unlagged and second lag of \code{Prey}, and unlagged \code{Salinity} by
specifying:

\code{lags = list(Predator = c(0:2), Prey = c(0, 2), Salinity = c(0))}.
}
\examples{
nt <- matrix(rep(1:30, 5), ncol = 5)
colnames(nt) <- c("A", "B", "C", "D", "E")
lags <- list(A = c(0, 1, 2), B = c(0, 1), C = c(0, 1, 2))
m1 <- pbsEDM(nt, lags)

nt <- data.frame(x = simple_ts)
lags <- list(x = 0:1)
m2 <- pbsEDM(nt, lags)

nt <- data.frame(x = simple_ts)
lags <- list(x = 0:1)
m3 <- pbsEDM(nt, lags, first_difference = TRUE)

}
\author{
Luke A. Rogers
}
