% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{pbsEDM_Evec}
\alias{pbsEDM_Evec}
\title{Do the `pbsEDM()` calculation for vector of `E` values}
\usage{
pbsEDM_Evec(N_t, E_vec = 1:10, ...)
}
\arguments{
\item{N_t}{[vector] Vector of non-differenced values $N_t$, with time assumed
to be `1:length(N_t)`.}

\item{E_vec}{The vector of embedding dimensions to try.}

\item{...}{Further options to pass to `pbsEDM()`.}
}
\value{
List of `pbsEDM` lists, each main component corresponds to a value
  of `E`, given by `results$E`
}
\description{
Only works for one variable, non-differenced $N_t$, for now, as I haven't thought how to incorporate
the lags variable properly yet.
}
\examples{
\donttest{
  aa <- pbsEDM_Evec(NY_lags_example$N_t)
}
}
\author{
Andrew Edwards
}
