% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{pbsLag}
\alias{pbsLag}
\title{Return a Lagged Matrix}
\usage{
pbsLag(x, n = 1)
}
\arguments{
\item{x}{[matrix()] A vector or column matrix}

\item{n}{[integer()] The lag sizes}
}
\value{
A matrix or vector. If `x` is a vector then returns a vector `length(x)`
  with `NA` for the first `n` values then the first `length(x) - n` values of
  `x`.
}
\description{
Return a Lagged Matrix
}
\examples{
pbsLag(1:10, 3)
pbsLag(matrix(rep(1:10, 2), nrow = 10), 3)
pbsLag(matrix(rep(1:10, 2), nrow = 10), c(3, 5))

pbsLag(matrix(rep(1:10, 2), nrow = 10), c(0, 1))
pbsLag(matrix(rep(1:10, 2), nrow = 10), c(0, 0))
pbsLag(matrix(rep(1:10, 2), nrow = 10), c(0, -1))

}
