% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{pbsLAG}
\alias{pbsLAG}
\title{Return a Lagged Matrix}
\usage{
pbsLAG(x, n)
}
\arguments{
\item{x}{[matrix()] A vector or column matrix}

\item{n}{[integer()] The lag sizes}
}
\value{
A matrix or vector
}
\description{
Return a Lagged Matrix
}
\examples{
pbsLAG(1:10, 3)
pbsLAG(matrix(rep(1:10, 2), nrow = 10), 3)
pbsLAG(matrix(rep(1:10, 2), nrow = 10), c(3, 5))

}
