% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{pbsSMAP}
\alias{pbsSMAP}
\title{Perform Out-of-Sample Forecasting via S-Mapping}
\usage{
pbsSMAP(
  nt,
  lags,
  local_weight = 0,
  forecast_distance = 1L,
  first_difference = FALSE,
  centre_and_scale = FALSE
)
}
\arguments{
\item{nt}{[data.frame()] A data frame with named columns}

\item{lags}{[list()] A named list of integer vectors}

\item{local_weight}{[numeric(1)]}

\item{forecast_distance}{[integer(1)]}

\item{first_difference}{[logical(1)]}

\item{centre_and_scale}{[logical(1)]}
}
\value{
A list
}
\description{
Perform Out-of-Sample Forecasting via S-Mapping
}
\details{
Only lags and columns explicitly named in `lags` are used. For
example, an unlagged time series in a column named `Salinity` could be 
included via `lags = list(..., Salinity = c(0))`.
}
\examples{
nt <- matrix(rep(1:30, 5), ncol = 5)
colnames(nt) <- c("A", "B", "C", "D", "E")
lags <- list(A = c(0, 1, 2), B = c(0, 1), C = c(0, 1, 2))
m1 <- pbsEDM(nt, lags)

nt <- data.frame(x = simple_ts)
lags <- list(x = 0:1)
m2 <- pbsSMAP(nt, lags)

}
\author{
Luke A. Rogers
}
