% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_explain_edm_movie}
\alias{plot_explain_edm_movie}
\title{Create movie to explain EDM}
\usage{
plot_explain_edm_movie(obj, ...)
}
\arguments{
\item{obj}{list object of class `pbsEDM`, an output from `pbsEDM()`}

\item{...}{extra values to use in calls to `plot_explain_edm()`; needs to include tstar}
}
\value{
movie (if used with gifski) to explain EDM; if combined with
  `par(mfrow=c(4,2))` (and not gifski) then will give multiple panels, but
  borders etc have not been set up properly for this and will need adjusting
}
\description{
Use with gifski in .Rmd - see `analyse_simple_time_series` vignette.
}
\examples{
\donttest{
  aa <- pbsEDM(NY_lags_example,
              lags = list(N_t = 0:1),
              first_difference = TRUE)
  plot_explain_edm_movie(aa, tstar = 15) # will only show last panel; see
                                         # `analyse_simple_time_series`
                                         #  vignette.
}
}
\author{
Andrew Edwards
}
