% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_observed}
\alias{plot_observed}
\title{Plot values of X(t), X(t-1), X(t-2) in one of various ways}
\usage{
plot_observed(
  obj,
  dim = 1,
  last.time.to.plot = NULL,
  max_time = NULL,
  late.col = "red",
  early.col = "black",
  early.col.lines = "lightgrey",
  late.num = 3,
  pt.type = "p",
  x.lab = expression("x"[t - 2]),
  y.lab = expression("x"[t - 1]),
  z.lab = expression("x"[t]),
  ...
)
}
\arguments{
\item{obj}{list object of class `pbsEDM`, an output from `pbsEDM()`}

\item{dim}{number of dimensions to plot: 1 for time series, 2 for X(t) vs X(t-2),
3 for X(t-2) v X(t-1) v X(t) - not now TODO}

\item{last.time.to.plot}{last time value of N[t] to use when plotting, so
final X[t] used is X[t-1] (since X[t] uses N[t+1])}

\item{max_time}{maximum time value for the time axis (will be different from `end` if
making a movie using Rmarkdown, since for that `end` will get iterated)}
}
\value{

}
\description{
Adapting from code within plotPanelMovie.df2. Assume raw data are one-dimensional
}
\author{
Andrew Edwards
}
