% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_pbsEDM_Evec}
\alias{plot_pbsEDM_Evec}
\title{Plot output from `pbsEDM_Evec()` as six panel plot, with data showns the
 same as for `plot.pbsEDM()` and sixth panel with predictions vs observation
 for each `E`}
\usage{
plot_pbsEDM_Evec(E_res, ...)
}
\arguments{
\item{E_res}{List of `pbsEDM` objects as output from `pbsEDM_Evec()`}

\item{...}{extra arguments to `plot.pbsEDM()`}
}
\value{
Six-panel figure in current plot environment
}
\description{
Plots time series of `N_t` and `Y_t`, 2d phase plots of `N_t` vs `N_{t-1}`
and `Y_t` vs `Y_{t-1}`, and 3d phase plot of `Y_t` vs `Y_{t-1}` vs
`Y_{t-2}`. Sixth panel shows predictions vs observations for different values
of `E`. See vignette "Analyse a simple time series" for full details.
}
\examples{
\donttest{
  aa_Evec <- pbsEDM_Evec(NY_lags_example$N_t)
  plot_pbsEDM_Evec(aa_Evec)

# For manuscript figures:
E_results <- pbsEDM_Evec(NY_lags_example$N_t)
postscript("six_panels_13.eps",   # 13 is last.time.to.plot
            height = 5.36,
            width = 9,
            horizontal=FALSE,
            paper="special")
plot_pbsEDM_Evec(E_results,
    last.time.to.plot = 13,
    portrait = FALSE)
dev.off()

postscript("six_panels_all.eps",
           height = 5.36,
           width = 9,
           horizontal=FALSE,
           paper="special")
plot_pbsEDM_Evec(E_results,
    last.time.to.plot = NULL, # to automatically plot all
    portrait = FALSE)
dev.off()
}
}
\author{
Andrew Edwards
}
