% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_pbsEDM_Evec}
\alias{plot_pbsEDM_Evec}
\title{Plot output from pbsEDM_Evec as six panel plot including final forecast v
 observed for each E}
\usage{
plot_pbsEDM_Evec(E_res, ...)
}
\arguments{
\item{E_res}{List of `pbsEDM` objects as output from `pbsEDM_Evec()`}
}
\value{
Six panel plot
}
\description{
<description>
}
\examples{
\donttest{
  aa <- pbsEDM_Evec(Nx_lags_orig$Nt)
  plot_pbsEDM_Evec(aa)
}
}
\author{
Andrew Edwards
}
