% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_pbsEDM_Evec}
\alias{plot_pbsEDM_Evec}
\title{Plot output from `pbsEDM_Evec()` as six panel plot, with data showns the
 same as for `plot.pbsEDM()` and sixth panel with predictions vs observation
 for each `E`}
\usage{
plot_pbsEDM_Evec(E_res, ...)
}
\arguments{
\item{E_res}{List of `pbsEDM` objects as output from `pbsEDM_Evec()`}

\item{...}{extra arguments to `plot.pbsEDM()`}
}
\value{
Six-panel figure in current plot environment
}
\description{
Plots time series of `N(t)` and `X(t)`, 2d phase plots of `N(t)` vs `N(t-1)`
and `X(t)` vs `X(t-1)`, and 3d phase plot of `X(t)` vs `X(t-1)` vs
`X(t-2)`. Sixth panel shows predictions vs observations for different values
of `E`. See vignette "Analyse a simple time series" for full details.
}
\examples{
\donttest{
  aa_Evec <- pbsEDM_Evec(Nx_lags_orig$Nt)
  plot_pbsEDM_Evec(aa_Evec)
}
}
\author{
Andrew Edwards
}
