% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_phase_2d}
\alias{plot_phase_2d}
\title{Plot 2d phase plot of `N_t` vs `N_{t-1}` or `Y_t` vs `Y_{t-1}`}
\usage{
plot_phase_2d(
  values,
  X.or.N = "X",
  par.mar.phase = c(3, 0, 1, 0),
  axis.range = NA,
  start = 1,
  last.time.to.plot = NULL,
  pt.type = "p",
  cobwebbing = TRUE,
  late.col = "red",
  early.col = "black",
  early.col.lines = "lightgrey",
  late.num = 3,
  label = NULL,
  label.cex = 0.7
)
}
\arguments{
\item{values}{vector of `N_t` or `Y_t` values}

\item{X.or.N}{"N" if raw non-differenced data, "X" for differenced data}

\item{par.mar.phase}{`par(mar)` values}

\item{axis.range}{range of axes, if NA then calculated from values}

\item{start}{first time step to plot (currently must be 1)}

\item{last.time.to.plot}{last time value of N[t] to use when plotting, so
final Y[t] used will be Y[t-1] (since Y[t] uses N[t+1])}

\item{pt.type}{`type` value for `points()`}

\item{cobwebbing}{if TRUE then add cobwebbing lines to phase plot}

\item{late.col}{colour in which to plot final `late.num` time steps}

\item{early.col}{colour in which to plot earlier time step points}

\item{early.col.lines}{colour in which to plot earlier time step points}

\item{late.num}{final number of `N[t]` time steps to plot in a different colour}

\item{label}{label to annotate plot, such as `(a)` etc. for six-panel figure}
}
\value{
Plots figure to current device
}
\description{
Shows the values in two-dimensional phase space (first differenced or not)
  with lag of 1. Cobwebbing shows how one point iterates to the next point
  in the time series.
}
\examples{
\donttest{
  plot_phase_2d(NY_lags_example$N_t, X.or.N = "N")
  plot_phase_2d(NY_lags_example$Y_t, X.or.N = "X")
}
}
\author{
Andrew Edwards
}
