% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_phase_3d}
\alias{plot_phase_3d}
\title{3d plot of x_t vs x_{t-1} vs x_{t-2}:}
\usage{
plot_phase_3d(
  obj,
  par.mgp.3d = c(3, 10, 0),
  par.mai.3d = c(0.1, 0.1, 0.1, 0.1),
  par.mar.3d = c(3, 0, 0, 0),
  x.lab = expression("X"[t - 2]),
  y.lab = expression("X"[t - 1]),
  z.lab = expression("X"[t]),
  last.time.to.plot = NULL,
  axis.range = NA,
  late.num = 3,
  pt.type = "p",
  late.col = "red",
  early.col = "black",
  early.col.lines = "lightgrey",
  axes.col = "darkblue",
  par.mar.phase = c(3, 0, 1, 0),
  par.mgp = c(1.5, 0.5, 0),
  ...
)
}
\arguments{
\item{obj}{`pbsEDM` object}
}
\value{

}
\description{
<description>
}
\author{
Andrew Edwards
}
