% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_pred_obs}
\alias{plot_pred_obs}
\title{Plot predictions versus observed values of `Y_t` for list of `pbsEDM`
objects, for various values of `E`}
\usage{
plot_pred_obs(
  E_res,
  E_components = 5,
  E_cols = c("orange", "blue", "green", "red", "black"),
  last.time.to.plot = NULL
)
}
\arguments{
\item{E_res}{A list of `pbsEDM` objects}

\item{E_components}{How many of the first E_res components to show}

\item{E_cols}{Vector of colours, one for each value of E}

\item{last.time.to.plot}{Last time value of `N_t` to use when plotting.}
}
\value{
Figure in the current plot enivironment
}
\description{
Plot the predicted versus observed values of `Y_t` for various values
of `E`, using the output (a list of `pbsEDM` objects) from `pbsEDM_Evec()`.
}
\examples{
\donttest{
  aa <- pbsEDM_Evec(NY_lags_example$N_t)
  plot_pred_obs(aa)
}
}
\author{
Andrew Edwards
}
