% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_pred_obs}
\alias{plot_pred_obs}
\title{Plot predictions of X[t] versus the observations for list of `pbsEDM` objects}
\usage{
plot_pred_obs(
  E_res,
  E_components = 5,
  E_cols = c("orange", "blue", "green", "red", "black"),
  last.time.to.plot = NULL,
  ...
)
}
\arguments{
\item{E_res}{A list of `pbsEDM` objects}

\item{E_components}{How many of the first E_res components to show}

\item{E_cols}{Vector of colours, one for each value of E}

\item{last.time.to.plot}{Last time value of `N[t]` to use when plotting.}

\item{...}{Extra arguments to pass to `plot()`}
}
\value{

}
\description{
<description>
}
\examples{
\donttest{
  aa <- pbsEDM_Evec(Nx_lags_orig$Nt)
  plot_pred_obs(aa)
}
}
\author{
Andrew Edwards
}
