% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_rho_Evec}
\alias{plot_rho_Evec}
\title{Plot rho versus E for output from pbsEDM_Evec
 <description>}
\usage{
plot_rho_Evec(E_res, ...)
}
\arguments{
\item{E_res}{List of `pbsEDM` objects as output from `pbsEDM_Evec()`}

\item{...}{Extra arguments to pass to `plot()`}
}
\value{
plot of rho versus E
}
\description{
Plot rho versus E for output from pbsEDM_Evec
 <description>
}
\examples{
\donttest{
  aa <- pbsEDM_Evec(Nx_lags_orig$Nt)
  plot_rho_Evec(aa)
}
}
\author{
Andrew Edwards
}
