% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_time_series}
\alias{plot_time_series}
\title{Plot the observed time series as either Nt or Xt}
\usage{
plot_time_series(
  values,
  X.or.N,
  par.mar.ts,
  t.axis.range = NULL,
  last.time.to.plot = NA,
  y.range,
  col.plot,
  col.plot.lines,
  pch.plot,
  start = 1,
  pt.type = "p"
)
}
\arguments{
\item{X.or.N}{"N" if raw non-differenced data, "X" for differenced data}

\item{iii}{time value to plot up to}
}
\value{

}
\description{
<description>
}
\author{
Andrew Edwards
}
