% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{psi_orig}
\alias{psi_orig}
\title{Psi values from a simple simulation model analysed using Andy's
original independent code}
\format{
Dataframe of size 100x4 with one row for each `tstar`, and columns:
  * `tstar`
  * `psi1`, `psi2`, `psi3`: values of $\psi_1, \psi_2$ and $\psi_3$ for that
  $t^*$.
}
\source{
Generated from Andy running `data-raw/Nx_lags.R`.
}
\usage{
psi_orig
}
\description{
For each $t^*$, contains values of $\psi_1, \psi_2$ and $\psi_3$ which are
the time indices of the nearest, second-nearest and third-nearest neighbours
in $\bf{x}$-space (vector $\bf{x}$, so in lagged difference space). There
are three nearest neighbours since embedding dimension $E=2$.
}
\keyword{datasets}
